back/requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-search-documents==11.5.2
backoff==2.2.1
bcrypt==4.2.0
blinker==1.7.0
build==1.2.2.post1
CacheControl==0.14.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docx==0.2.4
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.11
fastjsonschema==2.16.2
filelock==3.8.0
firebase-admin==5.3.0
Flask==2.2.2
flask-blueprints==0.0.6
Flask-Cors==3.0.10
Flask-Jsonpify==1.5.0
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
google-api-core==2.24.1
google-api-python-client==2.137.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.16.0
jws==0.1.3
kubernetes==32.0.1
langchain==0.2.17
langchain-community==0.2.4
langchain-core==0.2.43
langchain-openai==0.1.8
langchain-text-splitters==0.2.4
langsmith==0.1.147
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.0.8
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauth2client==3.0.0
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.82.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pillow==11.2.1
platformdirs==3.11.0
posthog==3.18.1
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.8.1
pydantic_core==2.33.2
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.1.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyrsistent==0.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.3.4
python-json-logger==2.0.4
python-jwt==2.0.1
python-multipart==0.0.20
pytz==2023.3
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
selenium==4.31.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
starlette==0.46.0
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.21.0
watchfiles==1.0.4
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0
Werkzeug==2.2.2
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0